Add server.js and fix posts validation
