+++ app.js
@@ -1,10 +1,5 @@
 const express = require('express');
 const app = express();
-const port = 3000;
-
-const db = require('./database');
-const routes = require('./rotas');
-
 const bodyParser = require('body-parser');
 
 app.use(
@@ -13,6 +8,4 @@ app.use(
   })
 );
 
-routes(app);
-
-app.listen(port, () => console.log(`App listening on port ${port}`));
+module.exports = app;
