+++ package.json
@@ -2,9 +2,9 @@
   "name": "blog",
   "version": "1.0.0",
   "description": "A simple blog made in Node.js",
-  "main": "app.js",
+  "main": "server.js",
   "scripts": {
-    "start": "nodemon app.js",
+    "start": "nodemon server.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
