+++ src/posts/posts-controlador.js
@@ -5,11 +5,6 @@ module.exports = {
   adiciona: (req, res) => {
     const post = new Post(req.body);
 
-    const erros = validationResult(req);
-    if (!erros.isEmpty()) {
-      return res.status(422).json({ erros: erros.array() });
-    }
-
     post
       .adiciona()
       .then(res.redirect('/posts'))
